fix(routing): move routes array to index.js
fix(routing): move routes array to index.js
