+++ src/common/routing/index.jsx
@@ -1,2 +1,92 @@
-export history from './history'
-export {getSidebarRoutes, getRouterRoutes, getRoutes} from './routes'
+// @flow
+import React from 'react'
+import {createBrowserHistory, createMemoryHistory} from 'history'
+import {asyncComponent} from 'react-async-component'
+import {Loader, Dimmer, Header, Icon} from 'semantic-ui-react'
+import _ from 'lodash'
+import Dashboard from 'containers/Dashboard'
+import Links from 'containers/Links'
+
+function asyncComponentCreator (url) {
+	return asyncComponent({
+		resolve: () => {
+			if (process.env.BROWSER) {
+				return require(`containers/${url}/index.jsx`).default
+			}
+			// flow-disable-next-line: The parameter passed to import() must be a literal string
+			return	import(/* webpackChunkName:"[index].[request]" */ `containers/${url}/index.jsx`)
+		},
+		LoadingComponent () {
+			return (
+				<Dimmer active>
+					<Loader size="large" active>
+						Loading page...
+					</Loader>
+				</Dimmer>
+			)
+		},
+		ErrorComponent () {
+			return (
+				<Dimmer active>
+					<Header inverted as="h2" icon textAlign="center">
+						<Icon name="refresh" />
+						Refresh
+						<Header.Subheader>Got error while loading page.</Header.Subheader>
+					</Header>
+				</Dimmer>
+			)
+		},
+		autoResolveES2015Default: true,
+		env: process.env.BROWSER ? 'browser' : 'node',
+		serverMode: 'resolve'
+	})
+}
+
+function routingFnCreator (useFor) {
+	const [AsyncNotFound] = ['NotFound'].map(asyncComponentCreator)
+	// Dashboard and Links included in build
+	// NotFound(404) is lazy
+	const routes: any[] = [
+		{
+			path: '/',
+			exact: true,
+			component: Dashboard,
+			name: 'Dashboard'
+		},
+		{
+			path: '/links',
+			exact: true,
+			component: Links,
+			name: 'Links'
+		},
+		{
+			component: AsyncNotFound,
+			name: '404'
+		}
+	]
+
+	const fns = {
+		// Returns routing for React-Router
+		routing () {
+			return routes
+				.map(a =>
+					_.pick(a, ['path', 'strict', 'exact', 'component'])
+				)
+		},
+		// Returns `name` + `path`. used in Header
+		meta () {
+			return routes
+				.map(a => _.pick(a, ['path', 'name']))
+		}
+	}
+
+	return fns[useFor]
+}
+
+const createRequiredHistory = process.env.BROWSER
+	? createBrowserHistory
+	: createMemoryHistory
+
+export const getMetaRoutes = routingFnCreator('meta')
+export const getRouterRoutes = routingFnCreator('routing')
+export const history = createRequiredHistory()
